Single-source schema primitives + round-trip-test the proto/domain boundary (closes #1502)
The proto↔domain conversions kept parallel copies of the same types in sync
by hand, so a field added to one copy but not the other was a silent data-drop
rather than a compile error. This replaces the hand-sync with single-sourced
types and property tests that fail when a field is dropped at the boundary.

## Dedupe the schema primitives
SchemaField, SchemaRelationship, EnumValue, EdgeField, RelationshipDirection,
RelationshipCardinality, and SchemaProtectionLevel were defined byte-for-byte
twice — once in nodespace-types (the wire shape shared with the Tauri client)
and once in nodespace-core. Deleted core's copies and re-exported the
nodespace-types definitions (core already depends on nodespace-types), moving
compute_edge_table_name and the primitive-level tests down with them. One
definition per type now; ~290 net lines of duplication removed.

The TaskNode/AiChatNode/SchemaNode struct pairs are intentionally NOT merged:
desktop-app/src-tauri is a thin gRPC client with no nodespace-core dependency,
so it can only reach the wire structs in nodespace-types. That split is
dependency-graph-forced, not debt, and stays.

## Round-trip property tests
- nodespace-core: TaskNode round trip — from_node(task.into_node()) recovers
  every typed field. This surfaced a real silent-drop bug: into_node wrote
  camelCase date keys (dueDate/startedAt/completedAt) while from_node and the
  persistence layer use snake_case, so the three date fields round-tripped to
  None. Fixed into_node to write the canonical snake_case storage keys; the
  wire converter already reads both, so the frontend response is unchanged.
- nodespace-types: Node → wire-JSON promotion contract — every field stored
  under properties.<type> is promoted to a top-level key (task/ai-chat/schema),
  with the nodespace:// uri injected.

proptest added as a dev-dependency; the recorded regression seed for the date
round-trip bug is checked in.
